we4x upadteeee/we4x-main/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Authlib==1.2.0
google-api-python-client==2.90.0
google-auth==2.22.0
python-dateutil==2.8.2
python-dotenv==1.0.0
Werkzeug<3.0.0
email-validator==2.1.1
gunicorn==21.2.0
psycopg2-binary
Flask-WTF==1.1.1
WTForms==3.0.1
